fix(*): reorganize project
