+++ src/components/Form/Chapter/ChapterTitleForm.js
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import Button from "./Button";
-import { db } from "../../../firebase";
+import { db } from "../../../lib/firebase";
 import { doc, getDoc, updateDoc } from "firebase/firestore";
 
 // Input component
