+++ src/components/Image/CloudinaryUploadWidget.js
@@ -1,5 +1,5 @@
 import { createContext, useEffect, useState } from "react";
-import { db } from "../../firebase";
+import { db } from "../../lib/firebase";
 import { doc, getDoc, updateDoc } from "firebase/firestore";
 
 const CloudinaryScriptContext = createContext();
