+++ src/components/VideoPlayer.js
@@ -1,8 +1,7 @@
 import React, { useState } from "react";
 import MuxPlayer from "@mux/mux-player-react";
-import { db } from "../firebase";
+import { db } from "../lib/firebase";
 import { doc, getDoc, updateDoc } from "firebase/firestore";
-import { cn } from "../utils/helperfunctions";
 
 const VideoPlayer = ({
   chapter,
