+++ src/hooks/useGetAllCategories.js
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react";
 import { collection, getDocs } from "firebase/firestore";
-import { db } from "../firebase";
+import { db } from "../lib/firebase";
 
 const useGetAllCategories = (setLoading) => {
   const [categories, setCategories] = useState([]);
