+++ src/hooks/useUserData.js
@@ -1,7 +1,7 @@
 import { useState, useEffect } from "react";
 import { useAuthState } from "react-firebase-hooks/auth";
 import { collection, query, where, getDocs } from "firebase/firestore";
-import { auth, db } from "../firebase";
+import { auth, db } from "../lib/firebase";
 
 const useUserData = () => {
   const [user, loading, error] = useAuthState(auth);
