+++ src/lib/firebase.jsx
@@ -3,7 +3,7 @@ import { GoogleAuthProvider, getAuth, signInWithPopup, signInWithEmailAndPasswor
 createUserWithEmailAndPassword, sendPasswordResetEmail, signOut } from "firebase/auth";
 import { getFirestore, query, getDocs, collection, where, addDoc } from "firebase/firestore";
 import { getStorage } from "firebase/storage";
-import { generateUsername } from "./utils/helperfunctions";
+import { generateUsername } from "../utils/helperfunctions";
 
 const firebaseConfig = {
   apiKey: process.env.REACT_APP_FIREBASE_API_KEY,
