+++ src/screens/Dashboard/AssignCourse.js
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 import { getDocs, collection } from "firebase/firestore";
 import imgPlaceholder from "./image-placeholder.png";
 import { assignCourse, selectCourse } from "../../utils/courseFunctions";
-import { db } from "../../firebase";
+import { db } from "../../lib/firebase";
 import Layout from "../../components/Dashboard/Layout";
 import { useNavigate } from "react-router-dom";
