+++ src/screens/Dashboard/Transactions.js
@@ -3,7 +3,7 @@ import { useNavigate, Link } from "react-router-dom";
 // import { Link } from "react-router-dom";
 import "./Transactions.css";
 import Layout from "../../components/Dashboard/Layout";
-import { db } from "../../firebase";
+import { db } from "../../lib/firebase";
 import { doc, getDoc } from "firebase/firestore";
 import { removeDuplicates } from "../../utils/helperfunctions";
