+++ src/utils/cartFunctions.js
@@ -1,6 +1,3 @@
-import { db } from "../firebase";
-import { v4 as uuidv4 } from "uuid";
-import { doc, serverTimestamp, setDoc, updateDoc } from "firebase/firestore";
 
 const addToCart = (props) => {
   const {
